Added some comments for groups and memberships, group avatar, minor bugfixes

+++ app/controllers/memberships_controller.rb
@@ -2,8 +2,12 @@ class MembershipsController < ApplicationController
 
   before_action :authenticate_user!
 
+  # Join a group
   def create
-    if params[:group_id] and params[:user_id]
+    puts params
+
+    unless params[:membership] and params[:membership][:group_id] and
+      params[:membership][:user_id]
       return render json: {
         status: 'error',
         errors: ["Missing param."]
@@ -11,7 +15,7 @@ class MembershipsController < ApplicationController
     end
 
     @user = current_user
-    @membership = @user.memberships.find_by(group_id: params[:group_id])
+    @membership = @user.memberships.find_by(group_id: params[:membership][:group_id])
     @membership_params = membership_create_params
     @membership_params[:join_date] = Date.current
 
@@ -34,7 +38,7 @@ class MembershipsController < ApplicationController
   def show
     @membership = Membership.find(params[:id])
     if @membership
-        render json: @membership        
+        render json: @membership
     else
       render json: {
         status: 'error',
@@ -47,8 +51,10 @@ class MembershipsController < ApplicationController
     @membership = Membership.find(params[:id])
     @user = current_user
 
-    if @membership 
+    if @membership
       if @user.is_super_user or @membership.is_admin
+
+        #Admins and super users can update memberships for every user
         if @membership.update(membership_update_params)
           render json: {
             status: 'success',
@@ -61,6 +67,7 @@ class MembershipsController < ApplicationController
           }, status: 403
         end
       else
+        #Normal user can only update own memberships
         if (@user.id == @membership.user_id) and @membership.update(membership_update_params)
           render json: {
             status: 'success',
@@ -71,41 +78,56 @@ class MembershipsController < ApplicationController
             status: 'error',
             errors: @membership.errors
           }, status: 403
-        end 
-      end       
-  
+        end
+      end
+
     else
       render json: {
         status: 'error',
         errors: ["membership not found"]
       }, status: 404
-    end  
+    end
 
   end
 
   def destroy
-
     @membership = Membership.find(params[:id])
-
+    @user = current_user
     if @membership
-
-      if @membership.destroy
-        render json: {
-          status: 'success'
-        }
+      if @user.is_super_user or @membership.is_admin
+        #Admins and super users can delete memberships for every user
+        if @membership.destroy
+          render json: {
+            status: 'success'
+          }
+        else
+          render json: {
+            status: 'error',
+            errors: @membership.errors
+          }, status: 403
+        end
       else
-        render json: {
-          status: 'error',
-          errors: @membership.errors
-        }, status: 403
+        #Normal users can only delete their own memberships
+        if (@user.id == @membership.user_id) and @membership.destroy
+          render json: {
+            status: 'success'
+          }
+        else
+          render json: {
+            status: 'error',
+            errors: @membership.errors
+          }, status: 403
+        end
       end
-
+    else
+      render json: {
+        status: 'error',
+        errors: ["membership not found"]
+      }, status: 404
     end
-
-
   end
 
-  private 
+  private
 
     def membership_create_params
       if @user.is_super_user or (@membership and @membership.is_admin)
@@ -113,15 +135,15 @@ class MembershipsController < ApplicationController
       else
         params.require(:membership).permit(:user_id, :group_id, :group_email_subscribed)
       end
-    end 
+    end
 
     def membership_update_params
-      
+
       if @user.is_super_user or (@membership and @membership.is_admin)
         params.require(:membership).permit(:is_admin, :group_email_subscribed, :position)
       else
         params.require(:membership).permit(:group_email_subscribed)
       end
-    end 
+    end
 
 end
